Liquibase + postgres

+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/service/GLIMAccountInfoWritePlatformServiceImpl.java
@@ -47,7 +47,7 @@ public class GLIMAccountInfoWritePlatformServiceImpl implements GLIMAccountInfoW
 
     @Override
     public void addGLIMAccountInfo(String accountNumber, Group group, BigDecimal principalAmount, Long childAccountsCount,
-            Boolean isAcceptingChild, Integer loanStatus, BigDecimal applicationId) {
+            Boolean isAcceptingChild, Integer loanStatus, Long applicationId) {
 
         GroupLoanIndividualMonitoringAccount glimAccountInfo = GroupLoanIndividualMonitoringAccount.getInstance(accountNumber, group,
                 principalAmount, childAccountsCount, isAcceptingChild, loanStatus, applicationId);
